[registers-cp] Fix stack usage problem
The stack was assigned to after the checkpoint_stack() call and before the
checkpoint_registers() call. This means that those stack variables are
not saved in the checkpoint. This is bad!

This was because on some compiler optimization levels, the
`checkpoint_restored_flag` address in memory was stored on the stack
because we set it before the registers_checkpoint, and check it after.
This is because a checkpoint_restore call can also end up at the
check...

So to remedy this, I moved the setting of the flag for both the
registers_checkpoint() and registers_restore() call into the
corresponding assembly functions.